blobman: Manifest files (#20)
- Reorganizes blobman shards so they are directories instead of a few
loose files.
- Each Shard is a directory, and the data for the shard is stored in
files, immediately beneath that directory in the filesystem.
- Child Shards (with longer prefixes) would be subdirectories beneath
other Shards, but individual Shards ignore their existence.
- Shards can be flushed individually, although that is not used
anywhere.
- The manifest file allows both the table and the pack to be flushed
concurrently, since the Manifest is only updated after a successful
flush.
- Manifests contain a Generation number, which will be used later to
compact the the Shard's data, when the table or pack gets too full.
- Manifest files contain 2 manifests, the even manifest and the odd
manifest. When loading the manifest we take whichever version has the
higher nonce.
- These changes remove the table header, that information is now encoded
in the manifest.